backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
